move store, upgrade inquirer-grouped-checkbox types
