+++ src/program/execute-deletions.ts
@@ -1,5 +1,5 @@
 import { green, red, yellow } from '../utils/colors.js'
-import store from './store.js'
+import store from './store/store.js'
 
 export async function executeDeletions(safe: string[], force: string[]): Promise<0 | 1> {
     store.setQueuedForDeletion(safe, force)
