+++ src/program/store/store.ts
@@ -1,4 +1,4 @@
-import { establishArgs } from '../utils/establish-args.js'
+import { establishArgs } from '../../utils/establish-args.js'
 import BranchStore from './BranchStore.js'
 
 const argv = establishArgs()
